Make ad_proc default to the function being public.  This decision was made by Don and Dave and a couple of others during the current bugbash.
Why not TIP it?  There's no semantic meaning to the switch, and ad_proc
doesn't require you to declare a procedure to be either -public or -private.
This was enforced in an automated test, but if we WERE to insist on it,
ad_proc should cause failure to declare a proc public or private and error.
That would be too harsh (would break a lot of client code, too).  Assuming
functions are public by default seems reasonable.  The only effect is to
change the behavior of the API browser if you've chosen private or public
viewing modes.